Fixed !help, added !alias, and properly handled amigious helper roles like physics.
Due to the number of aliases, !help was sending a message larger than
2000 messages. To fix this, I added !alias that will now return all the
aliases. Also, now if a user types a ping that might have multiple
roles like “!ping physics”, it will now tell the user the options that
they could use to get the ping to the right place.